Refactor: comment dto import를 index.ts로 수정
